Migrated code to follow set standards. Also some bugfixes :D
