+++ README.md
@@ -142,7 +142,7 @@ Cava is in [AUR](https://aur.archlinux.org/packages/cava/).
 
     yaourt -S cava
 
-The GUI branch (this) is also now in [AUR](https://aur.archlinux.org/packages/cava-gui-git/).
+The GUI branch (this) is also in [AUR](https://aur.archlinux.org/packages/cava-gui-git/).
 
     yaourt -S cava-gui-git
 
@@ -298,7 +298,7 @@ If cava quits unexpectedly or is force killed, echo must be turned on manually w
 | <kbd>up</kbd> / <kbd>down</kbd>| increase/decrease sensitivity |
 | <kbd>left</kbd> / <kbd>right</kbd>| increase/decrease bar width |
 | <kbd>a</kbd> / <kbd>s</kbd> | increase/decrease bar spacing |
-| <kbd>f</kbd> | toggle fullscreen (only in X and sdl modes) |
+| <kbd>f</kbd> | toggle fullscreen (only in window modes) |
 | <kbd>c</kbd> / <kbd>b</kbd>| change forground/background color |
 | <kbd>r</kbd> | Reload configuration |
 | <kbd>q</kbd> or <kbd>CTRL-C</kbd>| Quit C.A.V.A. |
@@ -339,7 +339,7 @@ $ pkill -USR1 cava
 
 ### GUI
 
-Cava (atleast the GUI branch) can now run in grapical modes.
+CAVA (this branch) can now run in grapical modes.
 
 NOTE: All of these options are in the config file (usually) at ~/.config/cava/config
 
@@ -359,47 +359,53 @@ Just a bit of explination. X11/Xlib utilizes hardware drawing (if availble), whi
 In the graphical modes you also have some other features, such as:
 
 
-Options that are unique to 'sdl' and 'x' modes alone:
+Options that are unique to 'sdl' and 'x':
+
+```
+Please use window category instead because of compatibility issues with the general category
+```
 
 Toggle fullscreen:
      
-    window_fullscreen = (1 for on and 0 for off)
+    fullscreen = (1 for on and 0 for off)
 
 Toggle window border:
     
-    window_border = (1 to enable and 0 to disable)
+    border = (1 to enable and 0 to disable)
 
-As a cross-compatibility issue, for 'x' and 'sdl' output modes please use:
+Change bar width/height (units are in pixels rather than characters):
     
-    win_bar_width = (specify value)
-    
-    win_bar_spacing = (specify value)
-
-instead of
-
     bar_width = (specify value)
-
-    bar_spacing = (specify value)
-
-for changing bar geometry, because 'sdl' and 'x' use pixel geometry rather than font based ones.
-
-Align the window to a part of the screen using:
     
-    window_alignment = 'value'
+    bar_spacing = (specify value)
+    
 
-Possible values are: 'top_left', 'top', 'top_right'. 'left', 'center', 'right', 'bottom_left', 'bottom', 'bottom_right' and 'none' if you don't want to position the window automaticly.
+
+Assign the window to a specific part of the screen by changing:
+    
+    alignment = 'value'
+
+Possible values are:
+
+```
+'top_left', 'top', 'top_right'. 'left', 'center', 'right', 'bottom_left', 'bottom', 'bottom_right' and 'none' if you don't want to position the window automaticly.
+```
 
 In addition to window aligment you can adjust the window using the following options:
     
-    window_x_padding = (specify value)
+    x_padding = (specify value)
     
-    window_y_padding = (specify value)
+    y_padding = (specify value)
 
-NOTE: These options don't apply if "window_aligment" is set to 'none'.
+NOTE: These options don't apply if "aligment" is set to 
+```
+'none'
+```
+.
 
-But there are features that are unique for the 'x' mode alone, and one of them is enabling the background to be transparent. However, you'll need a window composite manager running on your system (WARNING: generally causes slowdowns). This can be enabled by:
+But there are features that are unique for the 'x' mode alone, and one of them is enabling the window background to be transparent. However, you'll need a window composite manager running on your system (WARNING: generally causes slowdowns). This can be enabled by:
      
-     window_transparency = (0 disable, 1 enable)
+     transparency = (0 disable, 1 enable)
 
 A another feature is when you set the foreground color to 'default' in x mode. It will automaticly get the average desktop color and blend in with your background, looks really nice with transparency+no borders.
 To enable this you just have to change:
@@ -412,11 +418,13 @@ A screenshot with the 'default' foreground option with transparency+no borders (
 A quick demo showing off what can be done by enabling transparency and disabling window borders:
 ![transparency](http://i.imgur.com/QscuEh8.gif "transparency")
 
-NOTE: If you have issues with overlapping other windows, (fxp. if you click on the window, and it overlaps the previous one) you can change the following:
-    
-    window_keep_below = 1
+NOTE: To keep the window below any other (fxp. you want it to run in the background).
 
-However, this only works in X mode.
+You can use:
+    
+    keep_below = 1
+
+This only works in X mode, however.
 
 Thanks
 ------
