+++ output/graphical.h
@@ -1,0 +1,3 @@
+int w, h, windowX, windowY;
+unsigned char fs, borderFlag, transparentFlag, keepInBottom;
+char *windowAlignment;
